requirements.txt
# Core dependencies
aiogram==3.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0

# Search and matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# API integration
aiohttp==3.9.1
requests==2.31.0

# Internationalization
Babel==2.14.0

# Export functionality
openpyxl==3.1.2
reportlab==4.0.7

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development tools
black==23.12.0
isort==5.13.2
flake8==6.1.0